Updating avatar and badge to not return render method.
