+++ packages/react-avatar/src/components/Badge/Badge.tsx
@@ -4,17 +4,18 @@ import { useInlineTokens } from '@fluentui/react-theme-provider';
 import { useBadge } from './useBadge';
 import { BadgeProps } from './Badge.types';
 import * as classes from './Badge.scss';
+import { renderBadge } from './renderBadge';
 
 // Create a hook to resolve classnames.
 export const useBadgeClasses = makeClasses(classes);
 
 export const Badge = React.forwardRef((props: BadgeProps, ref: React.Ref<HTMLElement>) => {
-  const { render, state } = useBadge(props, ref);
+  const state = useBadge(props, ref);
 
   useBadgeClasses(state);
   useInlineTokens(state, '--badge');
 
-  return render(state);
+  return renderBadge(state);
 });
 
 Badge.displayName = 'Badge';
