+++ packages/react-avatar/src/components/Badge/index.ts
@@ -1,3 +1,4 @@
 export * from './Badge.types';
 export * from './useBadge';
 export * from './Badge';
+export * from './useBadge';
